Correct files to deploy
