+++ app/controllers/application_controller.rb
@@ -1,3 +1,2 @@
 class ApplicationController < ActionController::Base
-  skip_before_action :verify_authenticity_token
 end
