+++ app/views/layouts/application.html.erb
@@ -1,13 +1,14 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>RolChartsApi</title>
+    <title>RolCharts</title>
     <meta name="viewport" content="width=device-width,initial-scale=1">
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
 
     <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
+    <%= javascript_pack_tag 'index' %>
   </head>
 
   <body>
