+++ app/views/pages/index.html.erb
@@ -1,0 +1,1 @@
+<h1>ñadfjdasfasdfbjasñfasd</h1>
