--- config/initializers/cors.rb
@@ -1,6 +1,0 @@
-Rails.application.config.middleware.insert_before 0, Rack::Cors do
-  allow do
-    origins 'http://localhost:3000'
-    resource '*', headers: :any, methods: [:get, :post, :patch, :delete, :options, :head], credentials: true
-  end
-end
