--- cookies.txt
@@ -1,6 +1,0 @@
-# Netscape HTTP Cookie File
-# https://curl.haxx.se/docs/http-cookies.html
-# This file was generated by libcurl! Edit at your own risk.
-
-#HttpOnly_localhost	FALSE	/	FALSE	0	_rol_charts_app	pvdNBCabF0%2Bc1%2BLVEDPin4MKBVnN%2BDZsANGWgkjARiZLEgxyHRi%2FUu8IOo5i2u37oNR9DiHVAN5PNeRMe0G63MSC6qXrOoN7YDuuxSTGkwvro3jzgw8oj%2BVjkh7fVcfi37bWKO70CEkeL3xTGsn0UnNALX6W6gVP40iPnajHBgdtcNlVR%2Fp1sMrhNDqvwRMl3alxj%2BA%2B%2FQ%3D%3D--3LdjWpfV4uy6AaDl--nMi%2F%2BAn945TU7BjIHyLbaA%3D%3D
-#HttpOnly_localhost	FALSE	/	FALSE	0	__profilin	p%3Dt
