+++ package.json
@@ -1,11 +1,16 @@
 {
-  "name": "rol-charts-api",
+  "name": "rol-charts",
   "private": true,
   "dependencies": {
+    "@babel/preset-react": "^7.12.10",
     "@rails/actioncable": "^6.0.0",
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
     "@rails/webpacker": "5.2.1",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "prop-types": "^15.7.2",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "turbolinks": "^5.2.0"
   },
   "version": "0.1.0",
