--- test/fixtures/users.yml
@@ -1,11 +1,0 @@
-# Read about fixtures at https://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
-
-one:
-  username: MyString
-  email: MyString
-  password_digest: MyString
-
-two:
-  username: MyString
-  email: MyString
-  password_digest: MyString
