--- test/models/user_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class UserTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end
